+++ src/c++/Alignment.h
@@ -41,34 +41,34 @@ enum terminalGAPsStrategy {BOTH_PENALTIES, ONLY_GEP};
 class Alignment{
 public:
 	//To be used when it is already available a substitution matrix
-	Alignment(const char* fileName, const char* treeFileName, SubstitutionMatrix* s, float g, terminalGAPsStrategy tgs=ONLY_GEP);
+	Alignment(const char* fileName, const char* treeFileName, SubstitutionMatrix* s, double g, terminalGAPsStrategy tgs=ONLY_GEP);
 	//To be used when the substitution matrix has to be decided according with the alignment's average percentage of identity (only with proteins)
-	Alignment(const char* fileName, const char* treeFileName, const char* matrixSerie, float gep, float gop, terminalGAPsStrategy tgs = ONLY_GEP);
+	Alignment(const char* fileName, const char* treeFileName, const char* matrixSerie, double gep, double gop, terminalGAPsStrategy tgs = ONLY_GEP);
 
 	int getNumberOfSequences();
 	int getLength();
 	vector<GAP*>* getGAPS();
 	//Calculate the WSP-score (in the classic way, without using the counters)
-	float WSP();
+	double WSP();
 
 	terminalGAPsStrategy getTerminalGAPStrategy();
 
 	//They move a GAP of one position and return the improvement in the WSP-Score due to this change
-	float moveLeft(GAP* g);
-	float moveRight(GAP* g);
+	double moveLeft(GAP* g);
+	double moveRight(GAP* g);
 	/* Same as moveLeft and moveRight, but the improvement is not returned (not even calculated).
 	 * Used to do the backtracking and restore the cell as before the movement of the GAP g.*/
 	void goBackToLeft(GAP* g);
 	void goBackToRight(GAP* g);
 
 	//It returns the penalty of adding a GOP in the specified row (it depends on the weight of the row).
-	float getGOP(int row);
+	double getGOP(int row);
 
 	void save(const char* fileName);
 	virtual ~Alignment();
 protected:
 	//It calculates the mean of the identity scores of all couples of sequences in the alignment
-	float getAverageIdentityScore();
+	double getAverageIdentityScore();
 	//Given the file containing the guide tree, it generates the sequences' weights
 	void createWeights(const char* fileName);
 	//It reads sequences list (as a strings' vector) from a FASTA file
@@ -82,23 +82,23 @@ protected:
 	string* getNames();
 
 	//It calculates the identity score of two sequences (the percentage of identical residues found in the pairwise alignment)
-	float getIdentityScore(int firstRow, int secondRow);
+	double getIdentityScore(int firstRow, int secondRow);
 	//It calculate the distance between two sequences
-	float getPairwiseDistance(int firstRow, int secondRow);
+	double getPairwiseDistance(int firstRow, int secondRow);
 	//It returns the best substitution matrix for this alignment (based on average identity score)
-	void createSubstitutionMatrix(const char* matrixSerie, float GEP);
+	void createSubstitutionMatrix(const char* matrixSerie, double GEP);
 
 	/* Next two are used only to improve the readability of the code.
 	 * They give a simpler access to the alignment and counters matrices.*/
 	int& align(int row, int column);
-	float& counters(int row, int column);
+	double& counters(int row, int column);
 
 	//Score of two sequences in the specified rows (used by WSP). It requires also the number of GAPS inside the rows
-	float pairwise(int r1, int r2, int numberOfGAPSr1, int numberOfGAPSr2);
+	double pairwise(int r1, int r2, int numberOfGAPSr1, int numberOfGAPSr2);
 
 	/* It modify the character in position (row, column) and returns
 	 * the improvement obtained by changing it.*/
-	float changeCell(int row, int column, int newCharacter);
+	double changeCell(int row, int column, int newCharacter);
 	//Same as changeCell, but here the improvement is not calculated
 	void restoreCell(int row, int column, int newCharacter);
 
@@ -110,14 +110,14 @@ protected:
 	Alphabet* alphabet;
 	//Sequences name and properties (found in FASTA files)
 	vector<string> properties;
-	float* weights;
-	float weightsSUM;
+	double* weights;
+	double weightsSUM;
 	vector<GAP*>* GAPS;
 
-	float* countersMatrix;
+	double* countersMatrix;
 
 	//GAP opening penalty
-	float GOP;
+	double GOP;
 
 	terminalGAPsStrategy terminal;
 };
